chore: use schema to remove user
